test: rename test block for clarity
